Add chip damage. Fix signal to emit float value of health for more accurate health bar updates

+++ src/fight/Fight.cs
@@ -95,7 +95,7 @@ public partial class Fight : Node
         EmitSignal(SignalName.QuitRequested);
     }
 
-    private void CheckDeath(int from, int to)
+    private void CheckDeath(float from, float to)
     {
         if (to <= 0) EndFight();
     }
